updates by Skala
merged:

  gentoo-x86: refs/heads/gentoo.org
  party-overlay: 966963e6f0dac835b95a1dd0ad2ea54deec4a0f4
  squeezebox: cf7953988018d60d3638245b380406be494bd7ae
  funtoo-overlay: 0c055737025e305bdeb1ec42c3cef28bbf6a155d
  sabayon-for-gentoo: 2a1376e743936e026018a884692bc712049465e6

+++ sys-fs/btrfs-progs/btrfs-progs-9999.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2015 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-fs/btrfs-progs/btrfs-progs-9999.ebuild,v 1.39 2015/02/14 02:21:33 floppym Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-fs/btrfs-progs/btrfs-progs-9999.ebuild,v 1.42 2015/03/25 01:00:23 floppym Exp $
 
 EAPI=5
 
@@ -10,10 +10,11 @@ libbtrfs_soname=0
 
 if [[ ${PV} != 9999 ]]; then
 	MY_PV=v${PV}
-	KEYWORDS="~alpha ~amd64 ~arm ~ia64 ~mips ~ppc ~ppc64 ~sparc ~x86"
+	KEYWORDS="~alpha ~amd64 ~arm ~arm64 ~ia64 ~mips ~ppc ~ppc64 ~sparc ~x86"
 	SRC_URI="https://www.kernel.org/pub/linux/kernel/people/kdave/${PN}/${PN}-${MY_PV}.tar.xz"
 	S="${WORKDIR}"/${PN}-${MY_PV}
 else
+	WANT_LIBTOOL=none
 	inherit autotools git-r3
 	EGIT_REPO_URI="git://git.kernel.org/pub/scm/linux/kernel/git/kdave/btrfs-progs.git
 		https://git.kernel.org/pub/scm/linux/kernel/git/kdave/btrfs-progs.git"
@@ -24,17 +25,20 @@ HOMEPAGE="https://btrfs.wiki.kernel.org"
 
 LICENSE="GPL-2"
 SLOT="0/${libbtrfs_soname}"
-IUSE=""
+IUSE="+convert"
 
 RESTRICT=test # tries to mount repared filesystems
 
 RDEPEND="
 	dev-libs/lzo:2=
 	sys-libs/zlib:0=
-	sys-fs/e2fsprogs:0=
+	convert? (
+		sys-fs/e2fsprogs:0=
+		sys-libs/e2fsprogs-libs:0=
+	)
 "
 DEPEND="${RDEPEND}
-	sys-apps/acl
+	convert? ( sys-apps/acl )
 	app-text/asciidoc
 	app-text/docbook-xml-dtd:4.5
 	app-text/xmlto
@@ -49,10 +53,23 @@ src_prepare() {
 	if [[ ${PV} == 9999 ]]; then
 		eautoreconf
 		mkdir config || die
-		ln -s "${EPREFIX}"/usr/share/automake-${WANT_AUTOMAKE}/install-sh config/install-sh || die
+		local automakedir="$(autotools_run_tool --at-output automake --print-libdir)"
+		[[ -e ${automakedir} ]] || die "Could not locate automake directory"
+		ln -s "${automakedir}"/install-sh config/install-sh || die
 		ln -s "${EPREFIX}"/usr/share/gnuconfig/config.guess config/config.guess || die
 		ln -s "${EPREFIX}"/usr/share/gnuconfig/config.sub config/config.sub || die
 	fi
+}
+
+src_configure() {
+	local myeconfargs=(
+		--bindir="${EPREFIX}"/sbin
+
+		# --enable-convert is broken
+		# http://thread.gmane.org/gmane.comp.file-systems.btrfs/43873
+		$(usex convert '' --disable-convert)
+	)
+	econf "${myeconfargs[@]}"
 }
 
 src_compile() {
